requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
anyascii==0.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.1
django-allauth==0.51.0
django-bootstrap-modal-forms==2.2.0
django-bootstrap5==22.1
django-countries==7.4.2
django-countries-with-calling-codes==0.1.1
django-crispy-forms==1.14.0
django-daraja==1.2.0
django-filter==22.1
django-modelcluster==6.0
django-mpesa==2.0.14
django-permissionedforms==0.1
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-taggit==3.0.0
django-treebeard==4.5.1
django-utils-six==2.0
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
frozenlist==1.3.1
geoip2==4.6.0
html5lib==1.1
idna==3.4
l18n==2021.3
maxminddb==2.2.0
multidict==6.0.2
oauthlib==3.2.1
openpyxl==3.0.10
phonenumbers==8.12.56
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.2.1
telepath==0.3
types-cryptography==3.3.23
typing_extensions==4.4.0
urllib3==1.26.12
wagtail==4.0.2
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
yarl==1.8.1
